requirements.txt
setuptools~=50.3.2
Pillow~=8.0.1